fixing locale + naming refactor
