+++ src/module/activity/component/navigation/SignInRequestModal/SignInRequestModal.tsx
@@ -2,7 +2,6 @@ import { createBackdrop } from "@peersyst/react-native-components";
 import SignInRequestScreen, { SignInRequestScreenProps } from "module/activity/screen/SignInRequestScreen/SignInRequestScreen";
 import CardSelectModal, { CardSelectModalProps } from "module/common/component/feedback/CardSelectModal/CardSelectModal";
 import { useTranslate } from "module/common/hook/useTranslate";
-import { useState } from "react";
 
 export interface SignInRequestModalProps extends SignInRequestScreenProps, Omit<CardSelectModalProps, "title" | "dismissal" | "children"> {}
 
@@ -10,7 +9,7 @@ const SignInRequestModal = createBackdrop(({ signInRequest, ...modalProps }: Sig
     const translate = useTranslate();
 
     return (
-        <CardSelectModal title="Review connection" dismissal="close" style={{ height: "95%" }} {...modalProps}>
+        <CardSelectModal title={translate("reviewConnection")} dismissal="close" style={{ height: "95%" }} {...modalProps}>
             <SignInRequestScreen signInRequest={signInRequest} />
         </CardSelectModal>
     );
