+++ src/module/activity/types.ts
@@ -4,3 +4,10 @@ export interface ParsedPendingTransactions {
     title: string;
     data: CompleteTransactionRequestDto[];
 }
+
+export enum TransactionRequestStatus {
+    PENDING = "pending",
+    SIGNED = "signed",
+    DECLINED = "declined",
+    EXPIRED = "expired",
+}
